// src/interfaces/constantsInterface.ts
// Sidebar navigation
export interface NavData {
    title: string
    icon: string
    route: string
}
;[]

// Dashboard user metrics data interface
export interface MetricsData {
    icon: string
    data: string
    value: number
}
;[]

// Table header interface
export interface TableHeadings {
    id: number
    title: string
}
;[]